# src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResSeasonWithDetails": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"show": {
						"properties": {
							"title": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"episodes_count": {
						"type": "number",
						"format": "double"
					},
					"episodes": {
						"items": {
							"properties": {
								"thumbnail": {
									"type": "string"
								},
								"file": {
									"type": "string",
									"nullable": true
								},
								"episode_desc": {
									"type": "string"
								},
								"title": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"episodes_count",
					"episodes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResEpisodeN": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"episode_desc": {
						"type": "string"
					},
					"file": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResSeasonN": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"episodes_count": {
						"type": "number",
						"format": "double"
					},
					"episodes": {
						"items": {
							"$ref": "#/components/schemas/ResEpisodeN"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResSeason": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"show": {
						"properties": {
							"title": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"episodes_count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"episodes_count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqSeason": {
				"properties": {
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"show_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"show_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ReqSeason_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"season_no": {
						"type": "number",
						"format": "double"
					},
					"season_desc": {
						"type": "string"
					},
					"show_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ResEpisode": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"episode_desc": {
						"type": "string"
					},
					"file": {
						"type": "string",
						"nullable": true
					},
					"thumbnail": {
						"type": "string"
					},
					"season": {
						"properties": {
							"season_no": {
								"type": "number",
								"format": "double"
							},
							"title": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ReqEpisode_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"episode_desc": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"season_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Veridius-APIs",
		"version": "0.0.1",
		"description": "Veridius-APIs",
		"contact": {}
	},
	"paths": {
...
		"/seasons/getSeasonDet/{seasonId}": {
			"post": {
				"operationId": "GetSeasonById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResSeasonWithDetails"
								}
							}
						}
					}
				},
				"description": "Get season by ID with details",
				"summary": "Get season details",
				"tags": [
					"Seasons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "seasonId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/seasons/showDet/{showId}": {
			"get": {
				"operationId": "GetSeasonsByShow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResSeasonN"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all seasons by show ID",
				"summary": "Get seasons by show",
				"tags": [
					"Seasons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "showId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/seasons": {
			"post": {
				"operationId": "CreateSeason",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResSeason"
								}
							}
						}
					}
				},
				"description": "Create a new season",
				"summary": "Create season",
				"tags": [
					"Seasons"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqSeason"
							}
						}
					}
				}
			}
		},
		"/seasons/{seasonId}": {
			"put": {
				"operationId": "UpdateSeason",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResSeason"
								}
							}
						}
					}
				},
				"description": "Update season",
				"summary": "Update season",
				"tags": [
					"Seasons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "seasonId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ReqSeason_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSeason",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "string"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete season",
				"summary": "Delete season",
				"tags": [
					"Seasons"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "seasonId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/episodes/season/{seasonId}": {
			"get": {
				"operationId": "GetEpisodesBySeason",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalCount": {
											"type": "number",
											"format": "double"
										},
										"items": {
											"items": {
												"$ref": "#/components/schemas/ResEpisode"
											},
											"type": "array"
										}
									},
									"required": [
										"totalCount",
										"items"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all episodes by season ID",
				"summary": "Get episodes by season",
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "seasonId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/episodes/{episodeId}": {
			"get": {
				"operationId": "GetEpisodeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResEpisode"
								}
							}
						}
					}
				},
				"description": "Get episode by ID",
				"summary": "Get episode details",
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "episodeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResEpisode"
								}
							}
						}
					}
				},
				"description": "Update episode",
				"summary": "Update episode",
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "episodeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ReqEpisode_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "string"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete episode",
				"summary": "Delete episode",
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "episodeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/episodes/{seasonId}": {
			"post": {
				"operationId": "CreateEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResEpisode"
								}
							}
						}
					}
				},
				"description": "Create a new episode",
				"summary": "Create episode",
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "seasonId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"episode_desc": {
										"type": "string"
									},
									"thumbnail": {
										"type": "string"
									},
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}